move query interfaces into separated package
