+++ api/src/main/kotlin/github/com/hukuta94/delivery/api/adapter/http/order/OrderControllerV1.kt
@@ -3,7 +3,7 @@ package github.com.hukuta94.delivery.api.adapter.http.order
 import github.com.hukuta94.delivery.api.adapter.http.common.Location
 import github.com.hukuta94.delivery.core.application.usecase.order.CreateOrderCommand
 import github.com.hukuta94.delivery.core.application.usecase.order.CreateOrderUseCase
-import github.com.hukuta94.delivery.core.application.usecase.order.GetNotCompletedOrdersQuery
+import github.com.hukuta94.delivery.core.application.query.order.GetNotCompletedOrdersQuery
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
