+++ api/src/main/kotlin/github/com/hukuta94/delivery/api/startup/configuration/adapter/http/CourierControllerConfiguration.kt
@@ -1,7 +1,7 @@
 package github.com.hukuta94.delivery.api.startup.configuration.adapter.http
 
 import github.com.hukuta94.delivery.api.adapter.http.courier.CourierControllerV1
-import github.com.hukuta94.delivery.core.application.usecase.courier.GetBusyCouriersQuery
+import github.com.hukuta94.delivery.core.application.query.courier.GetBusyCouriersQuery
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
