+++ api/src/test/kotlin/github/com/hukuta94/delivery/api/adapter/http/courier/CourierControllerV1Test.kt
@@ -1,7 +1,7 @@
 package github.com.hukuta94.delivery.api.adapter.http.courier
 
-import github.com.hukuta94.delivery.core.application.usecase.courier.GetBusyCourierResponse
-import github.com.hukuta94.delivery.core.application.usecase.courier.GetBusyCouriersQuery
+import github.com.hukuta94.delivery.core.application.query.courier.GetBusyCourierResponse
+import github.com.hukuta94.delivery.core.application.query.courier.GetBusyCouriersQuery
 import github.com.hukuta94.delivery.core.domain.courier.newCourier
 import org.junit.jupiter.api.Test
 import org.mockito.BDDMockito.given
