+++ core/impl/src/main/kotlin/github/com/hukuta94/delivery/core/application/query/order/GetNotCompletedOrdersQuery.kt
@@ -1,4 +1,4 @@
-package github.com.hukuta94.delivery.core.application.usecase.order
+package github.com.hukuta94.delivery.core.application.query.order
 
 import java.util.*
